add login.py changed function cli() and put cli() to cli.py
